# lib/api-zod/src/generated/types/race.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * UltraRank API - ultrarunner rankings and race results
 * OpenAPI spec version: 0.1.0
 */
import type { RaceCategory } from './raceCategory';
import type { RaceStatus } from './raceStatus';
import type { RaceSurface } from './raceSurface';
import type { RaceWeatherConditions } from './raceWeatherConditions';

export interface Race {
  id: number;
  name: string;
  location: string;
  country: string;
  /** @nullable */
  countryCode?: string | null;
  date: string;
  distanceKm: number;
  /** @nullable */
  distanceLabel?: string | null;
  category: RaceCategory;
  surface: RaceSurface;
  /** @nullable */
  totalElevationM?: number | null;
  /** @nullable */
  description?: string | null;
  status: RaceStatus;
  /** @nullable */
  finishersCount?: number | null;
  /** @nullable */
  weatherConditions?: RaceWeatherConditions;
  /** @nullable */
  technicalityRating?: number | null;
  difficultyScore: number;
  /** @nullable */
  organizerId?: number | null;
  /** @nullable */
  organizerName?: string | null;
  createdAt?: string;
}